# docs/swagger.yaml
basePath: /
definitions:
  domain.CryptoWithdrawRequest:
    properties:
      amount:
        type: number
      asset_symbol:
        type: string
      network:
        type: string
      notes:
        type: string
      to_address:
        type: string
    required:
    - amount
    - asset_symbol
    - network
    - to_address
    type: object
  domain.CryptoWithdrawResponse:
    properties:
      amount:
        type: number
      asset_symbol:
        type: string
      status:
        type: string
      to_address:
        type: string
      transaction_id:
        type: string
      tx_hash:
        type: string
    type: object
...
  domain.DepositAddressResponse:
    properties:
      address:
        type: string
      asset_symbol:
        type: string
      network:
        type: string
    type: object
  domain.Disable2FARequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  domain.Enable2FAResponse:
    properties:
      qr_code_url:
        type: string
      secret:
        type: string
    type: object
  domain.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  domain.ExchangeRateResponse:
    properties:
      last_updated:
        type: string
      pair:
        type: string
      rate:
        type: number
    type: object
...
  domain.Login2FARequest:
    properties:
      code:
        type: string
      pre_auth_token:
        type: string
    required:
    - code
    - pre_auth_token
    type: object
  domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.LoginResponse:
    properties:
      expires_in:
        type: integer
      pre_auth_token:
        type: string
      token:
        type: string
      two_factor_required:
        type: boolean
    type: object
...
  domain.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  domain.RegisterResponse:
    properties:
      balances:
        items:
          $ref: '#/definitions/domain.WalletBalanceDTO'
        type: array
      email:
        type: string
      user_id:
        type: string
      username:
        type: string
      wallet_id:
        type: string
    type: object
  domain.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  domain.SwapRequest:
    properties:
      amount:
        type: number
      from_asset:
        type: string
      to_asset:
        type: string
    required:
    - amount
    - from_asset
    - to_asset
    type: object
  domain.SwapResponse:
    properties:
      fee_charged:
        type: number
      from_amount:
        type: number
      from_asset:
        type: string
      rate_used:
        type: number
      to_amount:
        type: number
      to_asset:
        type: string
      transaction_id:
        type: string
    type: object
  domain.TopUpRequest:
    properties:
      amount:
        type: number
      notes:
        type: string
    required:
    - amount
    type: object
  domain.TopUpResponse:
    properties:
      amount:
        type: number
      asset_symbol:
        type: string
      redirect_url:
        type: string
      snap_token:
        type: string
      transaction_id:
        type: string
      wallet_id:
        type: string
    type: object
...
  domain.Verify2FARequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
...
  domain.WithdrawFiatRequest:
    properties:
      account_name:
        type: string
      account_number:
        type: string
      amount:
        type: number
      bank_code:
        type: string
      notes:
        type: string
    required:
    - account_name
    - account_number
    - amount
    - bank_code
    type: object
  domain.WithdrawFiatResponse:
    properties:
      account_number:
        type: string
      admin_fee:
        type: number
      amount:
        type: number
      bank_code:
        type: string
      status:
        type: string
      total_deducted:
        type: number
      transaction_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: bintangridwan30@gmail.com
    name: Bintang Ridwan Pribadi
  description: |-
    E-Wallet REST API with user auth, top-up, transfer, and transaction history.
    Built with Go, Gin, GORM, PostgreSQL, JWT, and bcrypt.
  title: Ledger Backend API
  version: "1.0"
paths:
  /auth/2fa/disable:
    post:
      consumes:
      - application/json
      description: Disables 2FA by validating the current TOTP code.
      parameters:
      - description: Disable 2FA payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Disable2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: 2FA disabled successfully
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Invalid code
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Disable 2FA TOTP
      tags:
      - auth
  /auth/2fa/enable:
    post:
      description: Generates a new TOTP secret key and QR code provisioning URL for
        the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Secret generated successfully
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Enable2FAResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate 2FA TOTP secret key
      tags:
      - auth
  /auth/2fa/login:
    post:
      consumes:
      - application/json
      description: Verifies the TOTP code against the pre-auth token and returns the
        final JWT token.
      parameters:
      - description: Login 2FA challenge payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Login2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.LoginResponse'
              type: object
        "400":
          description: Invalid pre-auth token or code
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Invalid TOTP code
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Complete 2FA login challenge
      tags:
      - auth
  /auth/2fa/verify:
    post:
      consumes:
      - application/json
      description: Verifies the first TOTP code to confirm scanning and officially
        enables 2FA for the account.
      parameters:
      - description: Verify 2FA payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: 2FA enabled successfully
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Invalid code or signature
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Verify and enable 2FA TOTP
      tags:
      - auth
...
  /auth/google/callback:
    get:
      description: Handles the authorization code returned by Google, exchanges it
        for an access token, fetches the user's profile, and issues a JWT token.
      parameters:
      - description: OAuth state parameter
        in: query
        name: state
        required: true
        type: string
      - description: OAuth authorization code
        in: query
        name: code
        required: true
        type: string
      responses:
        "200":
          description: Google login successful
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.LoginResponse'
              type: object
        "400":
          description: Invalid oauth state or missing authorization code
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "502":
          description: 'Bad gateway: failed to retrieve Google profile'
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Google Login Callback
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token valid for 24 hours.
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.LoginResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and an associated wallet with balance
        0 in a single atomic transaction.
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.RegisterResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: Email or username already taken
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /crypto/address:
    get:
      description: Returns the authenticated user's EVM deposit address for the specified
        network and asset. Generates a new keypair if not exists.
      parameters:
      - description: Network name (polygon_amoy, sepolia)
        in: query
        name: network
        required: true
        type: string
      - description: Asset symbol (USDT, USDC)
        in: query
        name: asset_symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Address retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.DepositAddressResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get or generate crypto deposit address
      tags:
      - crypto
  /crypto/withdraw:
    post:
      consumes:
      - application/json
      description: Initiates an on-chain withdrawal of crypto to the specified destination
        address.
      parameters:
      - description: Withdrawal payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CryptoWithdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Withdrawal initiated
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.CryptoWithdrawResponse'
              type: object
        "400":
          description: Invalid input or address
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Insufficient balance
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Withdraw crypto to external address
      tags:
      - crypto
  /exchange/rate:
    get:
      description: Returns the current exchange rate for a given pair (e.g. USDT_IDR,
        USDC_IDR) from Binance API.
      parameters:
      - description: Pair symbol (e.g. USDT_IDR, USDC_IDR)
        in: query
        name: pair
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Exchange rate retrieved
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.ExchangeRateResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "502":
          description: External Binance service failure
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get exchange rate
      tags:
      - exchange
  /exchange/swap:
    post:
      consumes:
      - application/json
      description: Exchanges a specified amount of one asset to another asset internally.
        Applies a 0.5% platform fee.
      parameters:
      - description: Swap payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SwapRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Swap transaction completed
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.SwapResponse'
              type: object
        "400":
          description: Invalid input, same asset swap
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Insufficient balance
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Swap assets (Fiat/Crypto)
      tags:
      - exchange
  /fiat/withdraw:
    post:
      consumes:
      - application/json
      description: Initiates a withdrawal of Rupiah (IDR) to an external bank account
        via Midtrans Iris Sandbox. Deducts Rp 2.500 admin fee.
      parameters:
      - description: Withdrawal payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.WithdrawFiatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Withdrawal initiated
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.WithdrawFiatResponse'
              type: object
        "400":
          description: Invalid input, below minimum withdrawal
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Insufficient balance
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Withdraw Rupiah to bank account
      tags:
      - fiat
  /topup:
    post:
      consumes:
      - application/json
      description: Adds the specified amount to the authenticated user's wallet balance.
        Records a transaction with type 'topup'.
      parameters:
      - description: Top-up payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.TopUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Top-up successful
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.TopUpResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Top-up wallet balance
      tags:
      - wallet
  /transactions:
    get:
      description: Returns the authenticated user's transaction history with pagination
        and filters.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by asset (e.g. IDR, USDT, USDC)
        in: query
        name: asset
        type: string
      - description: Filter by transaction type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction history retrieved
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.TransactionHistoryResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get transaction history
      tags:
      - wallet
  /transfer:
    post:
      consumes:
      - application/json
      description: Transfers a specified amount from the authenticated user to another
        user. Uses pessimistic locking (SELECT ... FOR UPDATE) to prevent race conditions.
      parameters:
      - description: Transfer payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.TransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Transfer successful
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Invalid request, insufficient balance, or self-transfer
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Wallet not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Transfer money to another user
      tags:
      - wallet
  /wallet/dashboard:
    get:
      description: Returns the authenticated user's wallet balances (IDR, USDT, USDC)
        and estimated total IDR value.
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard data retrieved
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.DashboardResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get wallet dashboard
      tags:
      - wallet
  /webhooks/iris:
    post:
      consumes:
      - application/json
      description: Endpoint for Midtrans Iris disbursement notification webhook. Updates
        payout transaction status to success or failed (and refunds balance upon failure).
      parameters:
      - description: Midtrans Iris Notification Payload
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/domain.IrisCallbackItem'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Notification processed successfully
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Handle Midtrans Iris Callback Notification
      tags:
      - webhooks
  /webhooks/midtrans:
    post:
      consumes:
      - application/json
      description: Endpoint for Midtrans payment gateway notification webhook. Validates
        SHA-512 signature, processes billing status changes, and settles transaction
        balances.
      parameters:
      - description: Midtrans Notification Payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Notification processed successfully
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Invalid payload or signature validation failed
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Handle Midtrans Webhook Notification
      tags:
      - webhooks
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"